// Boardlight/src/api/generated/models/shopIndexLoloRequestsResponse.ts
/**
 * Generated by orval v6.17.0 🍺
 * Do not edit manually.
 * Blueboard
 * OpenAPI spec version: v4.0.0
 */

export interface ShopIndexLoloRequestsResponse {
  id?: number;
  title?: string | null;
  body?: string | null;
  userId?: string;
  acceptedAt?: string | null;
  deniedAt?: string | null;
  createdAt?: string;
  updatedAt?: string;
}